refactor(orchestration): improve naming clarity and organization
Address feedback to improve code clarity:
- Rename provisionPrerequisites → provisionNetwork (clearer purpose)
- Move GetNetworkID to network.go (belongs with network logic)
- Move computeProvisioner initialization to Reconcile() (clearer flow)
- Rename provisionResources → provisionImagesAndInfrastructure (specific)
- Rename provisionImagesAndIP → buildImagesAndFetchPublicIP (clearer)
- Rename provisionInfrastructure → provisionFirewallAndLoadBalancers (specific)

The naming now clearly distinguishes what each function does without confusion
between "resources" vs "infrastructure" terminology.

All tests passing.
