+++ internal/orchestration/network.go
@@ -3,29 +3,28 @@ package orchestration
 import (
 	"context"
 	"fmt"
-
-	"hcloud-k8s/internal/provisioning/compute"
 )
 
-// provisionPrerequisites handles initial setup and network provisioning.
-func (r *Reconciler) provisionPrerequisites(ctx context.Context) error {
+// provisionNetwork calculates subnets and provisions the cluster network.
+func (r *Reconciler) provisionNetwork(ctx context.Context) error {
 	// Calculate subnets
 	if err := r.config.CalculateSubnets(); err != nil {
 		return fmt.Errorf("failed to calculate subnets: %w", err)
 	}
 
-	// Provision network (must be first)
+	// Provision network
 	if err := r.infraProvisioner.ProvisionNetwork(ctx); err != nil {
 		return fmt.Errorf("failed to provision network: %w", err)
 	}
 
-	// Create compute provisioner with the provisioned network
-	r.computeProvisioner = compute.NewProvisioner(
-		r.infra,
-		r.talosGenerator,
-		r.config,
-		r.infraProvisioner.GetNetwork(),
-	)
-
 	return nil
+}
+
+// GetNetworkID returns the ID of the provisioned network.
+func (r *Reconciler) GetNetworkID() int64 {
+	network := r.infraProvisioner.GetNetwork()
+	if network == nil {
+		return 0
+	}
+	return network.ID
 }
